refactor: move data types to a separate file, rename and update imports
